Removed the use of DWORD in icarus code, and also changed bad uses of long to int - while there are no problems of using long in 64-bit builds for Windows (as they remain 32-bit in size), other OSs expand long to 64-bits under a 64-bit environment.

+++ code/icarus/BlockStream.cpp
@@ -126,15 +126,15 @@ int CBlockMember::ReadMember( char **stream, long *streamPos, CIcarus* icarus )
 	if ( m_id == CIcarus::ID_RANDOM )
 	{//special case, need to initialize this member's data to Q3_INFINITE so we can randomize the number only the first time random is checked when inside a wait
 		m_size = sizeof( float );
-		*streamPos += sizeof( long );
+		*streamPos += sizeof( int );
 		m_data = game->Malloc( m_size );
 		float infinite = game->MaxFloat();
 		memcpy( m_data, &infinite, m_size );
 	}
 	else
 	{
-		m_size = *(long *) (*stream + *streamPos);
-		*streamPos += sizeof( long );
+		m_size = *(int *) (*stream + *streamPos);
+		*streamPos += sizeof( int );
 		m_data = game->Malloc( m_size );
 		memcpy( m_data, (*stream + *streamPos), m_size );
 	}
